+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug fixes
 
 - Lack of OpenSSL support breaks some Node.js modules ([#2](https://github.com/anmonteiro/lumo/issues/2)).
+- Allow foreign libraries to side-effect the global scope when required.
 
 ## 1.0-alpha1 (2016-11-09)